Complete bubble sort task
Break line:  #indicates the end of one iteration sorting through the array
Next line: #skips to next iteration of loop if the
            current and next number are already in their correct positions
Before I had a break here, which exited the iteration of the array entirely,
instead of just skipping the current pair of numbers to compare the next pair
of numbers.

I did it! Let's go!!!